+++ source_codes/haskon/include/OdomImuLidar.h
@@ -6,22 +6,22 @@
 class OdomImuLidar : public Odom
 {
 public:
-    OdomImuLidar( int = 0, int = 0, int = 0 , int = 0, int = 0, int = 0 , int = 0, int = 0 );
+    OdomImuLidar( double = 0, double = 0, double = 0 , double = 0, double = 0, double = 0 , double = 0, double = 0 );
 
     // each class requires a public serialize function
     virtual void serialize( JSON::Adapter& );
 
     // setters
-    void setCovarianceXy( int );
-    void setCovarianceYaw( int );
+    void setCovarianceXy( double );
+    void setCovarianceYaw( double );
 
     // getters
     int getCovarianceXy() const;
     int getCovarianceYaw() const;
 
 private:
-    int covarianceXy;
-    int covarianceYaw;
+    double covarianceXy;
+    double covarianceYaw;
 
 };
